+++ src/Config.hs
@@ -239,6 +239,7 @@ module Config
   (
   -- * Parsing
     parse
+  , Position(..)
 
   -- * Pretty-printing
   , pretty
@@ -247,9 +248,10 @@ module Config
   , Section(..)
   , Value(..)
   , Atom(..)
+  , valueAnn
   ) where
 
-import Config.Value  (Atom(..), Value(..), Section(..))
+import Config.Value  (Atom(..), Value(..), Section(..), valueAnn)
 import Config.Parser (parseValue)
 import Config.Pretty (pretty)
 import Config.Lexer  (scanTokens)
@@ -263,10 +265,13 @@ import qualified Data.Text as Text
 
 -- | Parse a configuration file and return the result on the
 -- right, or the position of an error on the left.
+--
+-- The resulting value is annotated with source file locations.
+--
 -- Note: Text file lines are terminated by new-lines.
 parse ::
-  Text                {- ^ source text                    -} ->
-  Either String Value {- ^ error message or parsed value  -}
+  Text                           {- ^ source text                    -} ->
+  Either String (Value Position) {- ^ error message or parsed value  -}
 parse txt =
   case parseValue (layoutPass (scanTokens txt)) of
     Right x -> Right x
@@ -274,15 +279,15 @@ parse txt =
 
 explain :: Position -> Token -> String
 explain posn token
-   = show (posLine   posn) ++ ":"
-  ++ show (posColumn posn) ++ ": "
+   = "line "    ++ show (posLine   posn)
+  ++ " column " ++ show (posColumn posn) ++ ": "
   ++ case token of
        T.Error e     -> explainError e
-       T.Atom atom   -> "parse error: unexpected atom: " ++ Text.unpack atom
+       T.Atom atom   -> "parse error: unexpected atom: `" ++ Text.unpack atom ++ "`"
        T.String str  -> "parse error: unexpected string: " ++ show (Text.unpack str)
        T.Bullet      -> "parse error: unexpected bullet '*'"
        T.Comma       -> "parse error: unexpected comma ','"
-       T.Section s   -> "parse error: unexpected section: " ++ Text.unpack s
+       T.Section s   -> "parse error: unexpected section: `" ++ Text.unpack s ++ "`"
        T.Number 2  n -> "parse error: unexpected number: 0b" ++ showIntAtBase 2  intToDigit n ""
        T.Number 8  n -> "parse error: unexpected number: 0o" ++ showIntAtBase 8  intToDigit n ""
        T.Number 16 n -> "parse error: unexpected number: 0x" ++ showIntAtBase 16 intToDigit n ""