+++ src/Config/Lens.hs
@@ -9,6 +9,7 @@ module Config.Lens
   , list
   , values
   , sections
+  , ann
   ) where
 
 import Config.Value
@@ -26,53 +27,64 @@ import Data.Traversable
 key ::
   Applicative f =>
   Text {- ^ section name -} ->
-  (Value -> f Value) -> Value -> f Value
-key i f (Sections xs) = Sections <$> traverse (section i f) xs
-key _ _ v             = pure v
+  (Value a -> f (Value a)) -> Value a -> f (Value a)
+key i f (Sections a xs) = Sections a <$> traverse (section i f) xs
+key _ _ v               = pure v
 
 -- | Apply a function to the 'Value' contained inside the given
 -- 'Value' when it is a section name matches the given name.
 section ::
   Applicative f =>
   Text {- ^ section name -} ->
-  (Value -> f Value) -> Section -> f Section
-section i f s@(Section j v) | i == j = Section j <$> f v
-                            | otherwise = pure s
+  (Value a -> f (Value a)) -> Section a -> f (Section a)
+section i f s@(Section a j v) | i == j    = Section a j <$> f v
+                              | otherwise = pure s
 
 -- | Apply a function to the ['Section'] contained inside the given
 -- 'Value' when it is a @Sections@.
-sections :: Applicative f => ([Section] -> f [Section]) -> Value -> f Value
-sections f (Sections xs) = Sections <$> f xs
-sections _ v             = pure v
+sections :: Applicative f => ([Section a] -> f [Section a]) -> Value a -> f (Value a)
+sections f (Sections a xs) = Sections a <$> f xs
+sections _ v               = pure v
 
 -- | Apply a function to the 'Text' contained inside the given
 -- 'Value' when it is a @Text@.
-text :: Applicative f => (Text -> f Text) -> Value -> f Value
-text f (Text t) = Text <$> f t
-text _ v        = pure v
+text :: Applicative f => (Text -> f Text) -> Value a -> f (Value a)
+text f (Text a t) = Text a <$> f t
+text _ v          = pure v
 
 -- | Apply a function to the 'Text' contained inside the given
 -- 'Value' when it is a @Text@. This traversal is only valid
 -- if the output atom is a valid atom!
-atom :: Applicative f => (Atom -> f Atom) -> Value -> f Value
-atom f (Atom t) = Atom <$> f t
-atom _ v        = pure v
+atom :: Applicative f => (Atom -> f Atom) -> Value a -> f (Value a)
+atom f (Atom a t) = Atom a <$> f t
+atom _ v          = pure v
 
 -- | Apply a function to the 'Integer' contained inside the given
 -- 'Value' when it is a @Number@.
-number :: Applicative f => (Integer -> f Integer) -> Value -> f Value
-number f (Number b n) = Number b <$> f n
-number _ v            = pure v
+number :: Applicative f => (Integer -> f Integer) -> Value a -> f (Value a)
+number f (Number a b n) = Number a b <$> f n
+number _ v              = pure v
 
 -- | Apply a function to the ['Value'] contained inside the given
 -- 'Value' when it is a @List@.
-list :: Applicative f => ([Value] -> f [Value]) -> Value -> f Value
-list f (List xs) = List <$> f xs
-list _ v         = pure v
+list :: Applicative f => ([Value a] -> f [Value a]) -> Value a -> f (Value a)
+list f (List a xs) = List a <$> f xs
+list _ v           = pure v
 
 -- | Apply a function to the 'Value' elements inside the given
 -- 'Value' when it is a @List@.
 --
 -- > values = list . traverse
-values :: Applicative f => (Value -> f Value) -> Value -> f Value
+values :: Applicative f => (Value a -> f (Value a)) -> Value a -> f (Value a)
 values = list . traverse
+
+
+ann :: Functor f => (a -> f a) -> Value a -> f (Value a)
+ann f v =
+  case v of
+    Sections a x   -> (\a' -> Sections a' x  ) <$> f a
+    Number   a x y -> (\a' -> Number   a' x y) <$> f a
+    Floating a x y -> (\a' -> Floating a' x y) <$> f a
+    Text     a x   -> (\a' -> Text     a' x  ) <$> f a
+    Atom     a x   -> (\a' -> Atom     a' x  ) <$> f a
+    List     a x   -> (\a' -> List     a' x  ) <$> f a